requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
plotly==5.24.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
setuptools==75.1.0
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
xlrd==2.0.1
zipp==3.20.2
gunicornbeautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
xlrd==2.0.1
zipp==3.20.2
